Nothing but unused warnings now. Maybe it'll go now?
Will check in the AM.

+++ gplsrc/netfiles.c
@@ -6,18 +6,18 @@
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation; either version 2, or (at your option)
  * any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
- * 
+ *
  * Ladybridge Systems can be contacted via the www.openqm.com web site.
- * 
+ *
  * START-HISTORY:
  * 04 Oct 07  2.6-5 net_open() was not using ShortInt() when extracting file
  *                  number form response packet.
@@ -798,7 +798,7 @@ int net_scanindex(FILE_VAR *fvar, char *index_name, short int list_no,
 /* ======================================================================
    net_select()  -  Select file, returning id list                        */
 
-int net_select(FILE_VAR *fvar, STRING_CHUNK **str, long int *count) {
+int net_select(FILE_VAR *fvar, STRING_CHUNK **str, int32_t *count) {
   struct {
     short int fno;
   } ALIGN2 packet;
